+++ angular/projects/wcs-angular/src/lib/proxies.ts
@@ -1363,7 +1363,7 @@ import type { WcsTabChangeEvent as IWcsTabsWcsTabChangeEvent } from 'wcs-core';
 
 export declare interface WcsTabs extends Components.WcsTabs {
   /**
-   *
+   * 
 Emitted when the selected tab change.
    */
   tabChange: EventEmitter<CustomEvent<IWcsTabsWcsTabChangeEvent>>;
@@ -1371,7 +1371,7 @@ Emitted when the selected tab change.
 
 
 @ProxyCmp({
-  inputs: ['autoGrow', 'autocapitalize', 'autofocus', 'clearOnEdit', 'cols', 'debounce', 'disabled', 'enterkeyhint', 'icon', 'inputmode', 'maxlength', 'minlength', 'name', 'placeholder', 'readonly', 'required', 'rows', 'spellcheck', 'state', 'value', 'wrap'],
+  inputs: ['autoGrow', 'autocapitalize', 'autofocus', 'clearOnEdit', 'cols', 'debounce', 'disabled', 'enterkeyhint', 'icon', 'inputmode', 'maxlength', 'minlength', 'name', 'placeholder', 'readonly', 'required', 'resize', 'rows', 'spellcheck', 'state', 'value', 'wrap'],
   methods: ['fitContent', 'setFocus', 'getInputElement']
 })
 @Component({
@@ -1379,7 +1379,7 @@ Emitted when the selected tab change.
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
   // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
-  inputs: ['autoGrow', 'autocapitalize', 'autofocus', 'clearOnEdit', 'cols', 'debounce', 'disabled', 'enterkeyhint', 'icon', 'inputmode', 'maxlength', 'minlength', 'name', 'placeholder', 'readonly', 'required', 'rows', 'spellcheck', 'state', 'value', 'wrap'],
+  inputs: ['autoGrow', 'autocapitalize', 'autofocus', 'clearOnEdit', 'cols', 'debounce', 'disabled', 'enterkeyhint', 'icon', 'inputmode', 'maxlength', 'minlength', 'name', 'placeholder', 'readonly', 'required', 'resize', 'rows', 'spellcheck', 'state', 'value', 'wrap'],
 })
 export class WcsTextarea {
   protected el: HTMLElement;